support mono,xamarin studio

+++ samples/Samples/DeleteBucketSample.cs
@@ -59,9 +59,14 @@ namespace Aliyun.OSS.Samples
                     
                     nextMarker = result.NextMarker;
 
-                    client.DeleteObjects(new DeleteObjectsRequest(bucketName, keys));
-                    keys.Clear();
+					if (keys.Count != 0)
+					{
+                    	client.DeleteObjects(new DeleteObjectsRequest(bucketName, keys));
+                    	keys.Clear();
+					}
                 } while (result.IsTruncated);
+
+				client.DeleteBucket(bucketName);
 
                 Console.WriteLine("Delete bucket name:{0} succeeded ", bucketName);
             }
